Cleanup of Class/File/Test structure
Moved cluster specific resources to pf9/cluster/ (templates/*,
exceptions.py)
exceptions.py was broken down between core functionality and cluster
specific. Core Exceptions were moved to pf9/modules/exceptions.py
test/test_cli.py was broken out into the classes it is testing.
Basic test coverage was created for the 4 current cluster commands.
